rk3399: PCIe driver
Hi,

This diff adds support for the RK3399 PCIe controller to u-boot.  This
allows me to boot directly from NVMe on a NanoPC-T4, and hopefully also
soon on the Pinebook Pro.  The code is based on OpenBSD's RK3399 PCIe
driver, adjusted to the u-boot infrastructure.

I recently heard that there's some Radxa U-Boot fork that also has a
PCIe driver, but I did not have a look at this since at that point I
already had this diff.

I'm sure there are plenty of things that aren't U-Boot style or should
be improved, so it would be nice if someone could take over.  Thus, feel
free to take it, improve it, re-use it.

Best regards,
Patrick

Signed-off-by: Patrick Wildt <patrick@blueri.se>
Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>