spi: fu540: fix: use spi xfer bitlen for spi transfer
Use bitlen passed by dm_spi_ops rather than using spi-tx/
rx-bus-width from the device tree, to set the mode bits in
format register of spi controller present in FU540-C000 SoC
on HiFive Unleashed board.

This patch handles a case where controller mode in format
register (0x40) is configured as per the width specified
in the dt-node of the slave device. For instance if spi-tx-
bus-width and spi-rx-bus-width in the flash device node in
dt is set to 4 bit mode, the controller gets configured in
QUAD mode, whereas the spi nor scan tries to read the JEDEC
ID with the reg_proto set to SNOR_PROTO_1_1_1 and fails.

Signed-off-by: Sagar Shrikant Kadam <sagar.kadam@sifive.com>
Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>