roc-rk3399-pc: Configure the leds only during POR
ROC-RK3399-PC has specific set of configurations for on-board led setup.

Due to easiness for user to know the state of the board roc-rk339-pc board
code will setup the led on/off, and waiting for user to press power key.

All this needs to happen only during power-on-reset not for soft reset or WDT.
So add reset cause check to POR and configure leds.

Reported-by: Markus Reichl <m.reichl@fivetechno.de>
Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Tested-by: Suniel Mahesh <sunil@amarulasolutions.com>