spi: spi-mem: Xfer opcode alone for non spi-mem
Some of the SPI controllers have a special set of format
registers that defines how the transfer initiated to the
FIFO by means of I/O protocol lines.

Each mode of transfer from slave would be required to configure
the I/O protocol lines so-that the master would identify how
many number I/O protocol lines were used and alter the protocol
bits on the controller.

If a particular transfer combined opcode and address together,
then it would be difficult for the master to identify how many
I/O protocol lines are being used for opcode and address separately.

For example a transfer of SNOR_1_4_4 is not possible to identify
the master that how many I/O protocol lines by spi-nor if that
particular transfer initites the single xfer for both opcode and
address.

To address this issue (on these kind of SPI controllers) this
patch is trying to send the opcode alone as both opcode and
address send together in current code.

On the performance note there is no significant issue on the
transfer rate for adding this opcode xfer separately.

Sample test on 32MiB flash used in SiFive platform with
PP and RF opcodes.

With existing code:
=> sf update 0x90000000 0x0 0x2000000
device 0 whole chip
33554432 bytes written, 0 bytes skipped in 543.795s, speed 63191 B/s

With opcode alone xfer:
=> sf update 0x90000000 0x0 0x2000000
device 0 whole chip
33554432 bytes written, 0 bytes skipped in 541.739s, speed 63429 B/s

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>